backend/requirements.txt
altair==4.2.2
asgiref==3.8.1
attrs==25.3.0
bcrypt==4.1.2
blinker==1.9.0
cachetools==6.0.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
confluent-kafka==2.10.0
cryptography==45.0.3
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0.post1
entrypoints==0.4
et_xmlfile==2.0.0
fastapi==0.68.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.3
Flask-PyMongo==3.0.1
gitdb==4.0.12
GitPython==3.1.44
gunicorn==21.2.0
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numba==0.61.2
numpy==1.26.4
openpyxl==3.1.2
packaging==25.0
pandas==2.2.1
pillow==11.2.1
pluggy==1.6.0
protobuf==3.20.3
pyarrow==20.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.22
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.6.1
Pympler==1.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.5.0
pytz==2025.2
PyYAML==6.0.1
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
shap==0.41.0
six==1.17.0
slicer==0.0.7
smmap==5.0.2
starlette==0.14.2
streamlit==1.22.0
tenacity==8.5.0
threadpoolctl==3.6.0
toml==0.10.2
toolz==1.0.0
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.15.0
validators==0.35.0
Werkzeug==3.0.1
wheel==0.45.1
zipp==3.22.0
